+++ neurokit/__init__.py
@@ -1,7 +1,7 @@
 """
 neurokit module.
 """
-__version__ = "0.2.4"
+__version__ = "0.2.5"
 
 from .miscellaneous import *
 from .statistics import *