+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - `ecg_signal_quality`: Added rpeak parameter and returns an interpolated array of signal quality [#35](https://github.com/neuropsychology/NeuroKit.py/pull/35)
 - `eeg_complexity`: First attempt to compute complexity features of epochs (**since 0.2.0**) [#32](https://github.com/neuropsychology/NeuroKit.py/pull/32)
 - `emg_process`: Computes linear envelope and activation (**since 0.2.0**) [#32](https://github.com/neuropsychology/NeuroKit.py/pull/32)
-
+- `staircase`: Add a routine for staircase procedures used in psychophysics
 
 ### Major changes
 
@@ -24,7 +24,7 @@
 - Added [scripts](https://github.com/neuropsychology/NeuroKit.py/tree/master/utils/ecg_signal_quality_model_creation) for ECG signal quality model creation [#35](https://github.com/neuropsychology/NeuroKit.py/pull/35)
 - Started documenting ECG signal quality in docs [#35](https://github.com/neuropsychology/NeuroKit.py/pull/35)
 - Fixed bug in `ecg_hrv`
-
+- Moved `find_closest_in_list` and `find_following_duplicates` from miscellaenous to statistics
 ---------
 
 ## 0.1